winsvc: Delete - monitor services independently
This commit changes the Delete method to work like both Stop and Stop
whereby the function passed to iter() is responsible for waiting for the
service it is called with to complete the requested action before
returning.  Previously, the services where deleted then the Mgr would
keep iterating through the entire service list until it contained no
'vcap' services - this was not very efficient.

This commit also changes Delete to stop the service before deleting it.
Running services can be marked for deletion, but will not actually be
removed until they stop or the computer shutsdown.

There are also a couple of minor fixups included in this commit: bad
string format, remove unused loop counter and added two missing
comments.

[#148915345](https://www.pivotaltracker.com/story/show/148915345)